+++ crates/cli/tests/known_packages.rs
@@ -1,10 +1,15 @@
 use std::{
+    collections::BTreeSet,
     path::{Path, PathBuf},
     process::{Command, Stdio},
 };
 
 use url::Url;
 
+/// Download a WEBC package and make sure it would contain the expected
+/// libraries and commands.
+///
+/// We'll also try to generate the JavaScript bindings for good measure.
 macro_rules! codegen_test {
     (
         $( #[$meta:meta] )*
@@ -19,15 +24,20 @@ macro_rules! codegen_test {
             let temp = tempfile::tempdir().unwrap();
             let local_path = cached_url($url);
 
-            generate_bindings(&local_path, temp.path());
+            let meta = metadata(&local_path);
 
-            // TODO: check the libraries and commands were generated
+            // Make sure we detect the correct commands and libraries
+            insta::assert_display_snapshot!(format!("{meta:#}"));
+            assert_contains_libraries_and_commands(&meta, &[$($lib),*], &[$($cmd),*]);
+
+            // Make sure the binding generation doesn't fail
+            generate_bindings(&local_path, temp.path());
         }
     };
 }
 
 codegen_test! {
-    #[ignore]
+    #[ignore = "The webc file isn't being generated properly"]
     name: wabt,
     url: "https://registry-cdn.wapm.dev/packages/wasmer/wabt/wabt-1.0.33.webc",
     libraries: ["wabt"],
@@ -48,10 +58,51 @@ codegen_test! {
     commands: [],
 }
 
+fn assert_contains_libraries_and_commands(
+    meta: &serde_json::Value,
+    libraries: &[&str],
+    commands: &[&str],
+) {
+    let expected_libraries: BTreeSet<&str> = libraries.iter().copied().collect();
+    let actual_libraries: BTreeSet<&str> = meta["bindings"]
+        .as_array()
+        .unwrap()
+        .iter()
+        .map(|obj| obj["interface_name"].as_str().unwrap())
+        .collect();
+
+    assert_eq!(actual_libraries, expected_libraries);
+
+    let expected_commands: BTreeSet<&str> = commands.iter().copied().collect();
+    let actual_commands: BTreeSet<&str> = meta["commands"]
+        .as_array()
+        .unwrap()
+        .iter()
+        .map(|obj| obj["name"].as_str().unwrap())
+        .collect();
+
+    assert_eq!(actual_commands, expected_commands);
+}
+
+fn metadata(webc_file: &Path) -> serde_json::Value {
+    let mut cmd = Command::new(env!("CARGO_BIN_EXE_wit-pack"));
+    cmd.arg("show")
+        .arg("--format=json")
+        .arg(webc_file)
+        .stdin(Stdio::null())
+        .stderr(Stdio::piped())
+        .stdout(Stdio::piped());
+
+    let output = cmd.output().expect("Unable to invoke wit-pack");
+    assert_success(&output, &cmd);
+
+    serde_json::from_slice(&output.stdout).expect("Unable to deserialize the metadata")
+}
+
 fn generate_bindings(webc_file: &Path, out_dir: &Path) {
     let mut cmd = Command::new(env!("CARGO_BIN_EXE_wit-pack"));
     cmd.arg("js")
-        .arg(&webc_file)
+        .arg(webc_file)
         .arg("--out-dir")
         .arg(out_dir)
         .stdin(Stdio::null())
@@ -59,12 +110,17 @@ fn generate_bindings(webc_file: &Path, out_dir: &Path) {
         .stdout(Stdio::piped());
 
     let output = cmd.output().expect("Unable to invoke wit-pack");
+    assert_success(&output, &cmd);
+}
 
+fn assert_success(output: &std::process::Output, cmd: &Command) {
     if !output.status.success() {
         eprintln!("----- STDOUT -----");
         eprintln!("{}", String::from_utf8_lossy(&output.stdout).trim());
+
         eprintln!("----- STDERR -----");
         eprintln!("{}", String::from_utf8_lossy(&output.stderr).trim());
+
         panic!("Command failed: {cmd:?}");
     }
 }